+++ CHANGELOG.md
@@ -89,7 +89,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Updated packaging script for granule-invalidator to use `uv pip install` instead of `uv sync`.
   - Added `private_api_lambda_arn` output to the archive module and `private_api_lambda_arn` variable to the ingest module.
 - **CSD-91**
-  - Change update-granules-cmr-metadata-file-links task's updateGranuleIdentifiers flag from true to configurable environment variable `update_granule_identifiers`.
+  - Changed `update-granules-cmr-metadata-file-links` task config to accept a variable `updateGranuleIdentifiers` for whether or not to granule's `GranuleUr` value.
 - **CSD-85**
   - Changed `update-granules-cmr-metadata-file-links` task config to accept a variable `excludeDataGranule`
     for whether or not to add or update a `Granule.DataGranule` to the granule's metadata, for users who do not want one added or updated from what their granule metadata already is (defaults to `false`). See [update-granules-cmr-metadata-file-links](https://github.com/nasa/cumulus/tree/master/tasks/update-granules-cmr-metadata-file-links#readme) for more details.